+++ Tests/Main.fs
@@ -1,6 +1,7 @@
 namespace Tests
 
 open ResizeArrayT
+open Tests.Module3 // Add reference to Module3 tests
 
 module Main =
 
@@ -10,6 +11,7 @@ module Main =
     Mocha.runTests Tests.Extensions.tests |> ignore
     Mocha.runTests Tests.Module.tests |> ignore
     Mocha.runTests Tests.Module2.tests |> ignore
+    Mocha.runTests Tests.Module3.tests |> ignore // Add Module3 tests
 
     #else
 
@@ -21,6 +23,8 @@ module Main =
         runTestsWithCLIArgs [] [||] Tests.Module.tests
         |||
         runTestsWithCLIArgs [] [||] Tests.Module2.tests
+        |||
+        runTestsWithCLIArgs [] [||] Tests.Module3.tests // Add Module3 tests
 
 
     #endif