+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.25.0] - 2025-09-30
+### Added
+- add mapPrevNext function
+- add zipDefault function
+
 ## [0.24.0] - 2025-06-17
 ### Changed
 - equals does not fail for null input
@@ -62,7 +67,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - implementation ported from `Rarr` type in https://github.com/goswinr/FsEx/blob/main/Src/RarrModule.fs
 
-[Unreleased]: https://github.com/goswinr/ResizeArrayT/compare/v0.24.0...HEAD
+[Unreleased]: https://github.com/goswinr/ResizeArrayT/compare/v0.25.0...HEAD
+[0.25.0]: https://github.com/goswinr/ResizeArrayT/compare/v0.24.0...v0.25.0
 [0.24.0]: https://github.com/goswinr/ResizeArrayT/compare/v0.23.0...v0.24.0
 [0.23.0]: https://github.com/goswinr/ResizeArrayT/compare/v0.22.0...v0.23.0
 [0.22.0]: https://github.com/goswinr/ResizeArrayT/compare/v0.21.0...v0.22.0